pyproject.toml
[tool.poetry]
name = "misis-2526"
version = "0.1.0"
description = ""
authors = ["Афанасьев Максим Александрович"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
torch = "^2.6.0"
numpy = "^2.2.3"
pandas = "^2.2.3"
scikit-learn = "^1.6.1"
torchmetrics = "^1.6.1"
aim = "3.27.0"
plotly = "^6.0.0"
hyperopt = "^0.2.7"
matplotlib = "^3.10.1"
seaborn = "^0.13.2"
tokenizers = "^0.21.1"
pyarrow = "^19.0.1"
transformers = "^4.50.3"
accelerate = "^1.6.0"
torchdata = "^0.11.0"
datasets = "^3.5.0"
pytorch-metric-learning = "^2.8.1"
faiss-cpu = "^1.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
